# Minor changes to make integration easier.
# This commit fixes two problems  UIWebView & External .aar file.
# I have provided minor patch for it so please review and update it.

* I have replaced UIWebView with ui::Layout because iOS does not allow it
and rejects the app. Because of it we need to create two different projects for
Android (With UIWebView) and iOS (Without UIWebView).

* By default we have to use single GazLibrary-release.aar.aar for all our 
projects. But if we have to share that project with some other developer 
than path of GazLibrary-release.aar does not match and we have to update 
the path every time during project integration. That's why I updated the
code to add individual GazLibrary-release.aar file in every project.